refactor: separate logic from VerifyTemplate and VerifyForm
