+++ src/forms/VerifyForm.tsx
@@ -1,5 +1,8 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { useForm, SubmitHandler } from 'react-hook-form';
 import VerifyTemplate from './templates/VerifyTemplate';
+import { useVerificationMutation } from '@/services/Query/authorization/authorization.mutation';
+import { IAuthVerifyRequest } from '@/services/types';
 
 interface IVerifyTemplate {
   openPopup: boolean;
@@ -12,11 +15,36 @@ const VerifyForm = ({
   setOpenPopup,
   setIsVerify,
 }: IVerifyTemplate) => {
+  const {
+    register: verifyRegisterForm,
+    handleSubmit: verifyHandleSubmitForm,
+    reset: verifyResetForm,
+  } = useForm<IAuthVerifyRequest>();
+
+  const { mutateAsync: mutateVerifyEmail, isSuccess: verifyIsSuccess } =
+    useVerificationMutation();
+
+  useEffect(() => {
+    setIsVerify(verifyIsSuccess);
+  }, [setIsVerify, verifyIsSuccess]);
+
+  const requestVerify: SubmitHandler<IAuthVerifyRequest> = async (
+    activationCode
+  ) => {
+    mutateVerifyEmail(activationCode);
+    setOpenPopup(false);
+    verifyResetForm();
+  };
+
+  const onSubmitForm = () => {
+    verifyHandleSubmitForm(requestVerify)();
+  };
   return (
     <VerifyTemplate
       openPopup={openPopup}
       setOpenPopup={setOpenPopup}
-      setIsVerify={setIsVerify}
+      onSubmitForm={() => onSubmitForm()}
+      verifyRegisterForm={verifyRegisterForm}
     />
   );
 };
