+++ src/forms/templates/VerifyTemplate.tsx
@@ -1,42 +1,22 @@
-import Popup from '@/components/Popup';
-import { useVerificationMutation } from '@/services/Query/authorization/authorization.mutation';
-import { IAuthVerifyRequest } from '@/services/types';
+import React from 'react';
+import { UseFormRegister } from 'react-hook-form';
 import { TextField } from '@mui/material';
-import React, { useEffect } from 'react';
-import { useForm, SubmitHandler } from 'react-hook-form';
+import Popup from '@/components/Popup';
+import { IAuthVerifyRequest } from '@/services/types';
 
 interface IVerifyTemplate {
   openPopup: boolean;
   setOpenPopup: React.Dispatch<React.SetStateAction<boolean>>;
-  setIsVerify: React.Dispatch<React.SetStateAction<boolean>>;
+  onSubmitForm: () => void;
+  verifyRegisterForm: UseFormRegister<IAuthVerifyRequest>;
 }
 
 const VerifyTemplate = ({
   openPopup,
   setOpenPopup,
-  setIsVerify,
+  onSubmitForm,
+  verifyRegisterForm,
 }: IVerifyTemplate) => {
-  const {
-    register: verifyRegisterForm,
-    handleSubmit: verifyHandleSubmitForm,
-    reset: verifyResetForm,
-  } = useForm<IAuthVerifyRequest>();
-
-  const { mutateAsync: mutateVerifyEmail, isSuccess: verifyIsSuccess } =
-    useVerificationMutation();
-
-  useEffect(() => {
-    setIsVerify(verifyIsSuccess);
-  }, [setIsVerify, verifyIsSuccess]);
-
-  const requestVerify: SubmitHandler<IAuthVerifyRequest> = async (
-    activationCode
-  ) => {
-    mutateVerifyEmail(activationCode);
-    setOpenPopup(false);
-    verifyResetForm();
-  };
-
   return (
     <>
       <Popup
@@ -44,7 +24,7 @@ const VerifyTemplate = ({
         contentText="Please enter this verification code to get started on Twitter:"
         openPopup={openPopup}
         setOpenPopup={setOpenPopup}
-        onPopupSubmit={verifyHandleSubmitForm(requestVerify)}
+        onPopupSubmit={() => onSubmitForm()}
       >
         <TextField
           {...verifyRegisterForm('activationCode')}
